+++ CHANGELOG.md
@@ -47,3 +47,8 @@
     - Refactored CreateBrainForm by dividing it into modular form inputs.
     - Added GPT models from radio buttons by adding the chatModelList.ts utility function.
     - Renamed some components
+
+- **v-0.3.0**
+    - Refactored selectBrain.tsx to a component which renders links that use url searchParams instead of cookies.
+    - Removed components no longer necessary.
+